+++ lib/mapbox_search.dart
@@ -2,7 +2,7 @@ library mapbox_search;
 
 import 'dart:convert';
 import 'package:http/http.dart' as http;
-import 'package:color/color.dart';
+import 'package:mapbox_search/colors/color.dart';
 
 part 'src/location.dart';
 
@@ -15,3 +15,5 @@ part 'src/reverse_geocoding.dart';
 part 'src/places_search.dart';
 
 part "src/maki_icons.dart";
+
+part 'src/types.dart';